[FLINK-16201][sql] Support JSON_VALUE for blink planner
