+++ flink-table/flink-table-planner-blink/src/test/scala/org/apache/flink/table/planner/expressions/ScalarFunctionsTest.scala
@@ -4195,52 +4195,4 @@ class ScalarFunctionsTest extends ScalarTypesTestBase {
       "f55=f57",
       "true")
   }
-
-  //-------------------------------------------------------------------
-  // JSON functions
-  //-------------------------------------------------------------------
-  @Test
-  def testJsonValue(): Unit = {
-    // type casting test
-    testSqlApi("json_value('{\"foo\":100}', 'strict $.foo')", "100")
-    testSqlApi("json_value('{\"foo\":100}', 'strict $.foo' returning integer)", "100")
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo1' returning integer "
-      + "null on empty)", "null")
-    testSqlApi("json_value('{\"foo\":\"100\"}', 'strict $.foo1' returning boolean "
-      + "null on error)", "null")
-
-    // lax test
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo' null on empty)", "100")
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo' error on empty)", "100")
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo' default 'empty' on empty)", "100")
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo1' null on empty)", "null")
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo1' default 'empty' on empty)", "empty")
-    testSqlApi("json_value('{\"foo\":{}}', 'lax $.foo' null on empty)", "null")
-    testSqlApi("json_value('{\"foo\":{}}', 'lax $.foo' default 'empty' on empty)",
-      "empty")
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo' null on error)", "100")
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo' error on error)", "100")
-    testSqlApi("json_value('{\"foo\":100}', 'lax $.foo' default 'empty' on error)",
-      "100")
-
-    // path error test
-    testSqlApi("json_value('{\"foo\":100}', 'invalid $.foo' null on error)", "null")
-    testSqlApi("json_value('{\"foo\":100}', " + "'invalid $.foo' default 'empty' on error)",
-      "empty")
-
-    // strict test
-    testSqlApi("json_value('{\"foo\":100}', 'strict $.foo' null on empty)", "100")
-    testSqlApi("json_value('{\"foo\":100}', 'strict $.foo' error on empty)", "100")
-    testSqlApi("json_value('{\"foo\":100}', " + "'strict $.foo' default 'empty' on empty)",
-      "100")
-    testSqlApi("json_value('{\"foo\":100}', 'strict $.foo1' null on error)", "null")
-    testSqlApi("json_value('{\"foo\":100}', " + "'strict $.foo1' default 'empty' on error)",
-      "empty")
-    testSqlApi("json_value('{\"foo\":{}}', 'strict $.foo' null on error)", "null")
-    testSqlApi("json_value('{\"foo\":{}}', " + "'strict $.foo' default 'empty' on error)",
-      "empty")
-
-    // nulls
-    testSqlApi("json_value(cast(null as varchar), 'strict $')", "null")
-  }
 }
